+++ src/ansible_collections/solace/pubsub_plus/plugins/module_utils/solace_task_config.py
@@ -8,6 +8,7 @@ from ansible.module_utils.basic import AnsibleModule
 import ansible_collections.solace.pubsub_plus.plugins.module_utils.solace_sys as solace_sys
 from ansible_collections.solace.pubsub_plus.plugins.module_utils.solace_utils import SolaceUtils
 from ansible_collections.solace.pubsub_plus.plugins.module_utils.solace_error import SolaceInternalErrorAbstractMethod, SolaceInternalError, SolaceParamsValidationError
+import urllib.parse
 
 SOLACE_TASK_CONFIG_HAS_IMPORT_ERROR = False
 SOLACE_TASK_CONFIG_ERR_TRACEBACK = None
@@ -32,6 +33,9 @@ if not SOLACE_TASK_CONFIG_HAS_IMPORT_ERROR:
 class SolaceTaskConfig(object):
     def __init__(self, module: AnsibleModule):
         self.module = module
+
+    def get_module(self):
+        return self.module
 
     def validate_params(self):
         pass
@@ -136,6 +140,10 @@ class SolaceTaskBrokerConfig(SolaceTaskConfig):
 
     def is_solace_cloud(self) -> bool:
         return (self.solace_cloud_config is not None)
+
+    def get_broker_netloc(self):
+        parse_result = urllib.parse.urlparse(self.broker_url)
+        return parse_result.netloc
 
     def get_semp_url(self, path: str) -> str:
         return self.broker_url + path
@@ -257,10 +265,17 @@ class SolaceTaskBrokerConfig(SolaceTaskConfig):
         )
 
     @staticmethod
+    def _arg_spec_get_object_list_count():
+        return dict(
+            count=dict(type='int', default=-1, required=False)
+        )
+
+    @staticmethod
     def arg_spec_get_object_list_config_montor():
         d = dict(
             api=dict(type='str', default='config', choices=['config', 'monitor'])
         )
+        d.update(SolaceTaskBrokerConfig._arg_spec_get_object_list_count())
         d.update(SolaceTaskBrokerConfig._arg_spec_get_query_params())
         return d
 
@@ -269,6 +284,7 @@ class SolaceTaskBrokerConfig(SolaceTaskConfig):
         d = dict(
             api=dict(type='str', default='monitor', choices=['monitor'])
         )
+        d.update(SolaceTaskBrokerConfig._arg_spec_get_object_list_count())
         d.update(SolaceTaskBrokerConfig._arg_spec_get_query_params())
         return d