+++ src/ansible_collections/solace/pubsub_plus/plugins/modules/solace_service_authentication_ldap_profile.py
@@ -338,13 +338,14 @@ class SolaceServiceAuthenticationLdapProfileTask(SolaceBrokerCRUDTask):
         }
 
     def _make_solace_cloud_update_request(self, ldap_profile_name, settings):
+        module_op = SolaceTaskOps.OP_UPDATE_OBJECT
         # POST services/{service-id}/requests/ldapAuthenticationProfileRequests
         data = self.current_settings if self.current_settings else {}
         data.update(settings if settings else {})
         body = self._compose_solace_cloud_request_body(operation='update', settings=data)
         service_id = self.get_config().get_params()['solace_cloud_service_id']
         path_array = [SolaceCloudApi.API_BASE_PATH, SolaceCloudApi.API_SERVICES, service_id, SolaceCloudApi.API_REQUESTS, 'ldapAuthenticationProfileRequests']
-        return self.solace_cloud_api.make_service_post_request(self.get_config(), path_array, service_id, body)
+        return self.solace_cloud_api.make_service_post_request(self.get_config(), path_array, service_id, json_body=body, module_op=module_op)
 
     def _create_func_solace_cloud(self, ldap_profile_name, settings):
         data = {